+++ src/main/java/com/ssibssaggi/findex/controller/dto/IndexDataUpdateRequest.java
@@ -7,8 +7,8 @@ public record IndexDataUpdateRequest(
         Float lowPrice,
         Float versus,
         Float fluctuationRate,
-        Long tradingQuantity,     // Long으로 통일
-        Long tradingPrice,        // Long으로 통일
-        Long marketTotalAmount    // Long으로 통일
+        Integer tradingQuantity,     // Long으로 통일
+        Integer tradingPrice,        // Long으로 통일
+        Integer marketTotalAmount    // Long으로 통일
 ) {
 }
